+++ src/XRoadLib/Events/MessageEventArgs.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using XRoadLib.Serialization;
+
+namespace XRoadLib.Events
+{
+    public class MessageEventArgs : EventArgs
+    {
+        public XRoadMessage Message { get; }
+
+        public MessageEventArgs(XRoadMessage message)
+        {
+            Message = message;
+        }
+    }
+}