+++ src/XRoadLib/Serialization/ISerializerCache.cs
@@ -19,5 +19,7 @@ namespace XRoadLib.Serialization
         ITypeMap GetTypeMap(XmlQualifiedName qualifiedName, uint dtoVersion, IDictionary<XmlQualifiedName, ITypeMap> partialTypeMaps = null, bool undefined = false);
 
         XmlQualifiedName GetXmlTypeName(Type type);
+
+        bool IsSupportedProtocol(XRoadProtocol protocol);
     }
 }